+++ compiler/entry/jazz2tex.ml
@@ -1,29 +1,19 @@
-module J = Jasmin
-
-type arch = Amd64 | CortexM
+open Jasmin
+open Cmdliner
+open CommonCLI
 
 let parse_and_print arch call_conv =
-  let module A =
-    J.Arch_full.Arch_from_Core_arch
-      ((val match arch with
-            | Amd64 ->
-                (module (val J.CoreArchFactory.core_arch_x86 ~use_lea:false
-                               ~use_set0:false call_conv)
-                : J.Arch_full.Core_arch)
-            | CortexM ->
-                (module J.CoreArchFactory.Core_arch_ARM : J.Arch_full.Core_arch))) in
+  let module A = (val get_arch_module arch call_conv) in
   fun output file ->
-    let _, _, ast = J.Compile.parse_file A.arch_info file in
+    let _, _, ast = Compile.parse_file A.arch_info file in
     let out, close =
       match output with
       | None -> (stdout, ignore)
       | Some latexfile -> (open_out latexfile, close_out)
     in
     let fmt = Format.formatter_of_out_channel out in
-    Format.fprintf fmt "%a@." J.Latex_printer.pp_prog ast;
+    Format.fprintf fmt "%a@." Latex_printer.pp_prog ast;
     close out
-
-open Cmdliner
 
 let file =
   let doc = "The Jasmin source file to pretty-print" in
@@ -34,25 +24,6 @@ let output =
     "The file in which the result is written (instead of the standard output)"
   in
   Arg.(value & opt (some string) None & info [ "o"; "output" ] ~docv:"TEX" ~doc)
-
-let arch =
-  let alts = [ ("x86-64", Amd64); ("arm-m4", CortexM) ] in
-  let doc =
-    Format.asprintf "The target architecture (%s)" (Arg.doc_alts_enum alts)
-  in
-  let arch = Arg.enum alts in
-  Arg.(value & opt arch Amd64 & info [ "arch" ] ~doc)
-
-let call_conv =
-  let alts =
-    [ ("linux", J.Glob_options.Linux); ("windows", J.Glob_options.Windows) ]
-  in
-  let doc = Format.asprintf "Undocumented (%s)" (Arg.doc_alts_enum alts) in
-  let call_conv = Arg.enum alts in
-  Arg.(
-    value
-    & opt call_conv J.Glob_options.Linux
-    & info [ "call-conv"; "cc" ] ~docv:"OS" ~doc)
 
 let () =
   let doc = "Pretty-print Jasmin source programs into LATEX" in
@@ -65,7 +36,7 @@ let () =
     ]
   in
   let info =
-    Cmd.info "jazz2tex" ~version:J.Glob_options.version_string ~doc ~man
+    Cmd.info "jazz2tex" ~version:Glob_options.version_string ~doc ~man
   in
   Cmd.v info Term.(const parse_and_print $ arch $ call_conv $ output $ file)
   |> Cmd.eval |> exit